Add abstraction to get a timestamp, separate utils from constants
