+++ apps/hardhat/test/constants.ts
@@ -1,0 +1,25 @@
+import { Answer, Question } from '../types/Exam';
+
+const answers: Answer[] = [
+  {
+    answerText: 'aText1',
+    isCorrect: false,
+  },
+  {
+    answerText: 'aText2',
+    isCorrect: true,
+  },
+];
+
+export const MOCK_QUESTIONS: Question[] = [
+  {
+    header: 'questionHeader',
+    description: 'questionDescription',
+    answers,
+  },
+];
+
+export const EXAM_NAME = 'examName';
+export const EXAM_DESCRIPTION = 'examDescription';
+
+
